requirements.txt
python-gnupg==0.3.9
sphinx-me==0.3
django==1.10.4